Async copy and pasted from previous branch to this one (#5)
After importing AsyncStorage to Categories.js, useEffect along with a function to get the data from AsyncStorage is used when screen is loaded.

The items are added to AsyncStorage with the setItem function added in the addElement function so it only runs when the add button, in the pop-up modal, is clicked.

The items are removed in a similar fashion, having the removeItem function added in the removeElement function.